proper spilt header/source
