+++ co_sim_io/impl/file_serializer.hpp
@@ -15,10 +15,9 @@
 #define CO_SIM_IO_FILE_SERIALIZER_INCLUDED
 
 // System includes
-#include <iomanip>
 
 // Project includes
-#include "serializer.h"
+#include "serializer.hpp"
 
 namespace CoSimIO {
 
@@ -29,24 +28,7 @@ class FileSerializer : public Serializer
   public:
     ///this constructor simply wraps the standard Serializer and defines output to basic_iostream
     ///@param rTrace type of serialization to be employed
-    FileSerializer(std::string const& Filename, Serializer::TraceType const& rTrace=SERIALIZER_NO_TRACE)
-        : Serializer(nullptr, rTrace)
-    {
-        std::fstream* p_file = new std::fstream(std::string(Filename+".rest").c_str(), std::ios::binary|std::ios::in|std::ios::out);
-        if(!(*p_file))
-        {
-            delete p_file;
-            p_file = new std::fstream(std::string(Filename+".rest").c_str(), std::ios::binary|std::ios::out);
-        }
-        SetBuffer( p_file );
-        CO_SIM_IO_ERROR_IF_NOT(*pGetBuffer()) << "Error opening input file : "
-                                    << std::string(Filename+".rest") << std::endl;
-
-        // increase precision when using ascii
-        if (rTrace != SERIALIZER_NO_TRACE) {
-            *pGetBuffer() << std::setprecision(14);
-        }
-    }
+    FileSerializer(std::string const& Filename, Serializer::TraceType const& rTrace=SERIALIZER_NO_TRACE);
 
     /// Assignment operator.
     FileSerializer& operator=(FileSerializer const& rOther) = delete;
