+++ co_sim_io/impl/serializer.hpp
